requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
pydantic==2.6.1

pytest==8.0.0
httpx==0.26.0

pytest
httpx